requirements.txt
Flask==1.0.2
Pillow==6.1.0
numpy==1.17.0
torch==1.2.0
torchvision==0.4.0
